replaced example geno data with vcf file, consolidated code in variants.py, plot() functions now return ax object
